Prevent repeatedly build GET query params
Before this commit I forgot that the next link from response
has included the query string, so when we need to fetch next
resource it will repeatedly build the query string. Therefore
backend will receive a url with two same sets of query string.
This will cause error when query with project size.

To solve this we check if `status.link.next` is equal to
`status.link.root`, which is the root url without any query string.
if not equal, just request with `status.link.next`, else build query
string with filter.

Also I move the request building logic in reducer, in action we just
call `buildIssuesRequest` function in root reducer and pass url and
params to axios.