Build comparator query with JSON
Before this commit, we parse comparison query by:
1. parsing first character as operator
2. parsing the following string as value with a transformer function.
There's two problems:
1. It's not able to parse operator more than one character like '>='
2. It's not able to parse more than one comparison query at once.

So now, in browser we JSONized the comparison query with following
form: It will be an array of object, each object contains:
1. operator: the operator mapped to comparator, e.g. '>='
2. value: the value to compare, which is not string but the original
value.
In server side it will parse the JSON string, and build the WHERE
query object one element by one element. This is enough to handle
the query we need in this website. However this method cannot handle
logic relation like AND or OR.