Rewrite state.data.link in reducer
For the link information, we will need to know:
1. what's next to request?
2. Have we reached the end of query?
3. Have we read this page? (cache)
to build next request (or not to build, if we have read the last page).

To know these questions, we resolve the link header in response
in the actions/index.js, build link data with two attr:
1. isEnd
2. next
See the `buildLinkDataFromLinkHeader` function in actions/index.js.

And, if link header of that response is empty, that means the result
of query has only one or no page. So we can deduce that there's no
next page and we have reached the end.