Fix loading tip position
Make it on the center of viewport. Before this change we forgot
to consider the size of loading tip, so the loading tip will be
on the left bottom side of center.

It's difficult to fix it by `transform: translate(-0.5, -0.5)` since
the rotation animation. so we use `calc` to set the position val.