Handle data fetching error in action and reducer
Before this change, we didn't handle the case that there's no result.
The loading tip will keep in the screen since the `fetching` state
is still true. Now we handle the case, so if the client get no data
from server, we will set the fetching status so that the loading tip
will disappear.